backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
python-multipart
google-generativeai
